# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "API documentation generated with swagger-autogen",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/products/": {
      "get": {
        "description": "Get all products",
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "description": "Create a new product",
        "parameters": [
          {
            "in": "body",
            "name": "product",
            "description": "Product to create",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "price", "description", "category", "stock", "vendor"],
              "properties": {
                "name": { "type": "string" },
                "price": { "type": "number" },
                "description": { "type": "string" },
                "category": { "type": "string" },
                "stock": { "type": "integer" },
                "vendor": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Get a product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "description": "Update a product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "product",
            "description": "Product data to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "price": { "type": "number" },
                "description": { "type": "string" },
                "category": { "type": "string" },
                "stock": { "type": "integer" },
                "vendor": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "description": "Delete a product by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/vendors/": {
      "get": {
        "description": "Get all vendors",
        "responses": {
          "200": { "description": "OK" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "description": "Create a new vendor",
        "parameters": [
          {
            "in": "body",
            "name": "vendor",
            "description": "Vendor to create",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "address", "contact"],
              "properties": {
                "name": { "type": "string" },
                "address": { "type": "string" },
                "contact": {
                  "type": "object",
                  "required": ["email", "phone"],
                  "properties": {
                    "email": { "type": "string" },
                    "phone": { "type": "string" }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/vendors/{id}": {
      "get": {
        "description": "Get a vendor by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "description": "Update a vendor by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "in": "body",
            "name": "vendor",
            "description": "Vendor data to update",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "address": { "type": "string" },
                "contact": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "phone": { "type": "string" }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "description": "Delete a vendor by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}